UBUNTU: SAUCE: drm/i915/dp add a quirk for backlight issue
BugLink: https://bugs.launchpad.net/bugs/1923000

It looks like below commit in oem-5.10 introduces the backlight regression
   ade9f1d9f14ef drm/i915/dp: Program source OUI on eDP panels
But it's not the culprit as mainline kernel doesn't have this issue.
So, adding a quirk on this commit to make it only affect on a certain Dell
platform to avoid the issue temporarily. Will figure out the problem and
submit a formal fix later.

Signed-off-by: Chia-Lin Kao (AceLan) <acelan.kao@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>